Worked on power (.pow()).
