+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(calculator)
 
 set(CMAKE_CXX_STANDARD 11)
 
-add_executable(calulator arithmetic.cpp hugeint_math.cpp hugeint_cast.cpp hugeint.h hugeint_extra.cpp)
+add_executable(calulator arithmetic.cpp hugeint_math.cpp hugeint_cast.cpp hugeint.h)
