+++ arithmetic.cpp
@@ -205,4 +205,5 @@ int main () {
 	hugeint ans = function.getResult();
 	std::cout << "Answer: " << ans << std::endl;
 	std::cout << "Calculation time (s):" << global.reset() << std::endl;
+	std::cout << pow((hugeint)3, 7);
 }
